OneDrive/Desktop/filesharing/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
qrcode==7.4.2
Pillow==10.0.1
requests==2.31.0
